Alzheimer_Disease_Classifier-master/requirements.txt
dash~=3.3.0
dash-bootstrap-components~=2.0.4
imbalanced-learn~=0.14.0
pandas~=2.3.3
numpy~=2.3.5
scipy~=1.16.3
joblib~=1.5.2
scikit-learn==1.7.1
fpdf2~=2.8.5
protobuf~=5.29.5
plotly~=6.5.0
google-generativeai
gunicorn